excel_virtual_backend/excel_virtual_api/requirements.txt
asgiref==3.7.2
Django==4.1
djangorestframework==3.14.0
mysqlclient==2.2.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
